HOME PAGE START
I just added some content to the top of the ccdebts index page and was
going to build it out as the main page. But, I'm going to start building
another main index page now. on this page will be the user content and
summary of information. its controller will also calculate the cost of
debt.